move SwapERC20 helper to utils to avoid typechain confusion (#1208)
