+++ source/pool/package.json
@@ -30,7 +30,7 @@
   "devDependencies": {
     "@airswap/constants": "^4.1.0",
     "@airswap/types": "^4.1.1",
-    "@airswap/utils": "^4.1.2",
+    "@airswap/utils": "^4.1.3",
     "prompt-confirm": "^2.0.4"
   },
   "publishConfig": {
