+++ source/registry/package.json
@@ -32,7 +32,7 @@
   },
   "devDependencies": {
     "@airswap/constants": "^4.1.0",
-    "@airswap/utils": "^4.1.2",
+    "@airswap/utils": "^4.1.3",
     "prompt-confirm": "^2.0.4"
   }
 }
