+++ source/swap-erc20/tsconfig.json
@@ -3,5 +3,5 @@
   "compilerOptions": {
     "outDir": "./typechain"
   },
-  "files": ["./typechain/index.ts", "./events.ts"]
+  "files": ["./typechain/index.ts"]
 }
