+++ source/wrapper/package.json
@@ -27,12 +27,12 @@
     "verify": "hardhat run ./scripts/verify.js"
   },
   "dependencies": {
-    "@airswap/swap-erc20": "4.1.1",
+    "@airswap/swap-erc20": "4.1.2",
     "@openzeppelin/contracts": "^4.8.3"
   },
   "devDependencies": {
     "@airswap/constants": "^4.1.0",
-    "@airswap/utils": "^4.1.2",
+    "@airswap/utils": "^4.1.3",
     "@airswap/types": "^4.1.1",
     "@uniswap/v2-periphery": "^1.1.0-beta.0",
     "prompt-confirm": "^2.0.4"
