+++ tools/utils/index.ts
@@ -1,6 +1,7 @@
-import { ethers, BigNumber as BigNumberEthers } from 'ethers'
 import * as url from 'url'
+import { ethers, BigNumber as BigNumberEthers } from 'ethers'
 import { explorerUrls } from '@airswap/constants'
+import { SwapERC20, FullSwapERC20 } from '@airswap/types'
 
 export * from './src/pricing'
 export * from './src/swap'
@@ -12,6 +13,79 @@ export function getReceiptUrl(chainId: number, hash: string): string {
 
 export function getAccountUrl(chainId: number, address: string): string {
   return `${explorerUrls[chainId]}/address/${address}`
+}
+
+const parseTransfer = (log: any, tokenInterface: ethers.utils.Interface) => {
+  let parsed
+  let transfer
+  try {
+    parsed = tokenInterface.parseLog(log)
+    if (parsed.name === 'Transfer')
+      transfer = {
+        token: log.address,
+        from: parsed.args[0],
+        to: parsed.args[1],
+        amount: ethers.BigNumber.from(parsed.args[2]),
+      }
+  } catch (e) {
+    return null
+  }
+  return transfer
+}
+
+export const getFullSwapERC20 = async (
+  swapInterface: ethers.utils.Interface,
+  tokenInterface: ethers.utils.Interface,
+  tx: ethers.providers.TransactionResponse,
+  event: SwapERC20
+): Promise<FullSwapERC20> => {
+  const receipt = await tx.wait()
+  const transfers: any = []
+  for (let i = 0; i < receipt.logs.length; i++) {
+    let parsed: ethers.utils.LogDescription
+    try {
+      parsed = swapInterface.parseLog(receipt.logs[i])
+    } catch (e) {
+      continue
+    }
+    if (parsed && parsed.name === 'SwapERC20') {
+      let transfer: any
+      while (i--) {
+        if ((transfer = parseTransfer(receipt.logs[i], tokenInterface))) {
+          transfers.push(transfer)
+        }
+      }
+      break
+    }
+  }
+
+  const [fee, signer, sender] = transfers
+  if (fee.from !== event.signerWallet)
+    throw new Error(
+      'unable to get SwapERC20 params: found incorrect fee transfer (wrong signerWallet)'
+    )
+  if (signer.from !== event.signerWallet)
+    throw new Error(
+      'unable to get SwapERC20 params: found incorrect signer transfer (wrong signerWallet)'
+    )
+  if (signer.from !== sender.to)
+    throw new Error(
+      'unable to get SwapERC20 params: signer transfer mismatched sender transfer'
+    )
+  if (sender.from !== signer.to)
+    throw new Error(
+      'unable to get SwapERC20 params: sender transfer mismatched signer transfer'
+    )
+
+  return {
+    ...event,
+    signerToken: signer.token,
+    signerAmount: signer.amount.toString(),
+    senderWallet: sender.from,
+    senderToken: sender.token,
+    senderAmount: sender.amount.toString(),
+    feeAmount: fee.amount.toString(),
+  }
 }
 
 export function checkResultToErrors(
