+++ tools/utils/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@airswap/utils",
-  "version": "4.1.2",
+  "version": "4.1.3",
   "description": "AirSwap: Utilities for Developers",
   "repository": {
     "type": "git",
